Massively reorganize LR1 modules
Move all of the lr1 specific data structures into `core`, which means I
can make them pub while keeping the lr1/mod interface fairly clean.
