+++ lalrpop/src/lr1/core/mod.rs
@@ -1,184 +1,186 @@
-//! Core LR(1) state construction algorithm.
+//! Core LR(1) types.
 
 use kernel_set;
-use session::Session;
 use grammar::repr::*;
-use lr1::{Action, Conflict, Item, Items, State, StateIndex, TableConstructionError};
-use lr1::first;
-use lr1::lookahead::Lookahead;
+use std::fmt::{Debug, Formatter, Error};
 use std::rc::Rc;
-use util::{map, Multimap, Set};
+use util::{Map, Prefix};
 
-#[cfg(test)] mod test;
+use super::lookahead::Lookahead;
 
-pub fn build_lr1_states<'grammar>(session: &Session,
-                                  grammar: &'grammar Grammar,
-                                  start: NonterminalString)
-                                  -> Result<Vec<State<'grammar>>,
-                                            TableConstructionError<'grammar>>
-{
-    let lr1 = LR1::new(session, grammar);
-    lr1.build_states(start)
+#[derive(Copy, Clone, Hash, PartialEq, Eq, PartialOrd, Ord)]
+pub struct Item<'grammar> {
+    pub production: &'grammar Production,
+    /// the dot comes before `index`, so `index` would be 1 for X = A (*) B C
+    pub index: usize,
+    pub lookahead: Lookahead,
 }
 
-struct LR1<'session, 'grammar> {
-    session: &'session Session,
-    grammar: &'grammar Grammar,
-    first_sets: first::FirstSets,
-}
-
-impl<'session, 'grammar> LR1<'session, 'grammar> {
-    fn new(session: &'session Session, grammar: &'grammar Grammar) -> Self {
-        LR1 {
-            session: session,
-            grammar: grammar,
-            first_sets: first::FirstSets::new(grammar),
-        }
+impl<'grammar> Item<'grammar> {
+    pub fn prefix(&self) -> &[Symbol] {
+        &self.production.symbols[..self.index]
     }
 
-    fn build_states(&self, start_nt: NonterminalString)
-                    -> Result<Vec<State<'grammar>>, TableConstructionError<'grammar>>
-    {
-        let mut kernel_set = kernel_set::KernelSet::new();
-        let mut states = vec![];
-        let mut errors = 0;
+    pub fn to_lr0(&self) -> LR0Item<'grammar> {
+        LR0Item { production: self.production, index: self.index }
+    }
 
-        // create the starting state
-        kernel_set.add_state(
-            self.transitive_closure(
-                self.items(start_nt, 0, Lookahead::EOF)));
+    pub fn can_shift(&self) -> bool {
+        self.index < self.production.symbols.len()
+    }
 
-        while let Some(items) = kernel_set.next() {
-            let index = StateIndex(states.len());
+    pub fn can_reduce(&self) -> bool {
+        self.index == self.production.symbols.len()
+    }
 
-            if index.0 % 10000 == 0 && index.0 > 0 {
-                log!(self.session, Verbose, "{} states created so far.", index.0);
-            }
-
-            let mut this_state = State { index: index, items: items.clone(),
-                                         tokens: map(), gotos: map(),
-                                         conflicts: map() };
-
-            // group the items that we can transition into by shifting
-            // over a term or nonterm
-            let transitions: Multimap<Symbol, Item<'grammar>> =
-                items.vec
-                     .iter()
-                     .filter_map(|item| item.shifted_item())
-                     .collect();
-
-            for (symbol, items) in transitions.into_iter() {
-                let items = self.transitive_closure(items);
-                let next_state = kernel_set.add_state(items);
-
-                match symbol {
-                    Symbol::Terminal(s) => {
-                        let action = Action::Shift(next_state);
-                        let prev = this_state.tokens.insert(Lookahead::Terminal(s), action);
-                        assert!(prev.is_none()); // cannot have a shift/shift conflict
-                    }
-
-                    Symbol::Nonterminal(s) => {
-                        let prev = this_state.gotos.insert(s, next_state);
-                        assert!(prev.is_none());
-                    }
-                }
-            }
-
-            // finally, consider the reductions
-            for item in items.vec.iter().filter(|i| i.can_reduce()) {
-                let action = Action::Reduce(item.production);
-                let prev = this_state.tokens.insert(item.lookahead, action);
-                if let Some(conflict) = prev {
-                    log!(self.session, Verbose, "Encountered conflict in state {}",
-                         index.0);
-                    this_state.conflicts.entry(item.lookahead)
-                                        .or_insert(vec![])
-                                        .push(Conflict {
-                                            state: index,
-                                            production: item.production,
-                                            action: conflict,
-                                        });
-                    errors += 1;
-                }
-            }
-
-            // extract a new state
-            states.push(this_state);
-
-            if self.session.stop_after(errors) {
-                log!(self.session, Verbose,
-                     "{} conflicts encountered, stopping.", errors);
-                break;
-            }
-        }
-
-        if states.iter().any(|s| !s.conflicts.is_empty()) {
-            Err(TableConstructionError { states: states })
+    pub fn shifted_item(&self) -> Option<(Symbol, Item<'grammar>)> {
+        if self.can_shift() {
+            Some((self.production.symbols[self.index],
+                  Item { production: self.production,
+                         index: self.index + 1,
+                         lookahead: self.lookahead }))
         } else {
-            Ok(states)
+            None
         }
     }
 
-    fn items(&self,
-             id: NonterminalString,
-             index: usize,
-             lookahead: Lookahead)
-             -> Vec<Item<'grammar>>
-    {
-        self.grammar.productions_for(id)
-                    .iter()
-                    .map(|production| {
-                        debug_assert!(index <= production.symbols.len());
-                        Item { production: production,
-                                        index: index,
-                                        lookahead: lookahead }
-                    })
-                    .collect()
-    }
-
-    // expands `state` with epsilon moves
-    fn transitive_closure(&self, mut items: Vec<Item<'grammar>>)
-                          -> Items<'grammar>
-    {
-        let mut counter = 0;
-
-        let mut set: Set<Item<'grammar>> =
-            items.iter().cloned().collect();
-
-        while counter < items.len() {
-            let new_items: Vec<_> =
-                items[counter..]
-                .iter()
-                .filter_map(|item| {
-                    let shift_symbol = item.shift_symbol();
-                    match shift_symbol {
-                        None => None, // requires a reduce
-                        Some((Symbol::Terminal(_), _)) => None, // requires a shift
-                        Some((Symbol::Nonterminal(nt), remainder)) => {
-                            Some((nt, remainder, item.lookahead))
-                        }
-                    }
-                })
-                .flat_map(|(nt, remainder, lookahead)| {
-                    let (first_set, _) =
-                        self.first_sets.first(self.grammar, remainder, lookahead);
-                    first_set.iter(self.grammar)
-                             .flat_map(|l| self.items(nt, 0, l))
-                             .collect::<Vec<_>>()
-                })
-                .filter(|&item| set.insert(item))
-                .collect();
-
-            counter = items.len();
-            items.extend(new_items);
+    pub fn shift_symbol(&self) -> Option<(Symbol, &[Symbol])> {
+        if self.can_shift() {
+            Some((self.production.symbols[self.index], &self.production.symbols[self.index+1..]))
+        } else {
+            None
         }
-
-        items.sort();
-        items.dedup();
-
-        Items { vec: Rc::new(items) }
     }
 }
 
+#[derive(Copy, Clone, PartialEq, Eq, PartialOrd, Ord, Hash)]
+pub struct LR0Item<'grammar> {
+    pub production: &'grammar Production,
+    pub index: usize
+}
+
+impl<'grammar> LR0Item<'grammar> {
+    pub fn prefix(&self) -> &[Symbol] {
+        &self.production.symbols[..self.index]
+    }
+}
+
+#[derive(Copy, Clone, Hash, PartialEq, Eq, PartialOrd, Ord)]
+pub struct StateIndex(pub usize);
+
+#[derive(Clone, Debug, Hash, PartialEq, Eq, PartialOrd, Ord)]
+pub struct Items<'grammar> {
+    pub vec: Rc<Vec<Item<'grammar>>>
+}
+
+impl<'grammar> kernel_set::Kernel for Items<'grammar> {
+    type Index = StateIndex;
+
+    fn index(c: usize) -> StateIndex {
+        StateIndex(c)
+    }
+}
+
+#[derive(Debug)]
+pub struct State<'grammar> {
+    pub index: StateIndex,
+    pub items: Items<'grammar>,
+    pub tokens: Map<Lookahead, Action<'grammar>>,
+    pub conflicts: Map<Lookahead, Vec<Conflict<'grammar>>>,
+    pub gotos: Map<NonterminalString, StateIndex>,
+}
+
+#[derive(Copy, Clone, Debug, PartialEq, Eq)]
+pub enum Action<'grammar> {
+    Shift(StateIndex),
+    Reduce(&'grammar Production),
+}
+
+#[derive(Copy, Clone, Debug, PartialEq, Eq)]
+pub struct Conflict<'grammar> {
+    // when in this state...
+    pub state: StateIndex,
+
+    // we can reduce...
+    pub production: &'grammar Production,
+
+    // but we can also...
+    pub action: Action<'grammar>,
+}
+
+#[derive(Debug)]
+pub struct TableConstructionError<'grammar> {
+    // LR(1) state set. Some of these states are in error.
+    pub states: Vec<State<'grammar>>,
+}
+
+impl<'grammar> Debug for Item<'grammar> {
+    fn fmt(&self, fmt: &mut Formatter) -> Result<(), Error> {
+        write!(fmt, "{} ={} (*){} [{:?}]",
+               self.production.nonterminal,
+               Prefix(" ", &self.production.symbols[..self.index]),
+               Prefix(" ", &self.production.symbols[self.index..]),
+               self.lookahead)
+    }
+}
+
+impl<'grammar> Debug for LR0Item<'grammar> {
+    fn fmt(&self, fmt: &mut Formatter) -> Result<(), Error> {
+        write!(fmt, "{} ={} (*){}",
+               self.production.nonterminal,
+               Prefix(" ", &self.production.symbols[..self.index]),
+               Prefix(" ", &self.production.symbols[self.index..]))
+    }
+}
+
+impl Debug for Lookahead {
+    fn fmt(&self, fmt: &mut Formatter) -> Result<(), Error> {
+        match *self {
+            Lookahead::EOF => write!(fmt, "EOF"),
+            Lookahead::Terminal(s) => write!(fmt, "{}", s),
+        }
+    }
+}
+
+impl Debug for StateIndex {
+    fn fmt(&self, fmt: &mut Formatter) -> Result<(), Error> {
+        write!(fmt, "S{}", self.0)
+    }
+}
+
+impl<'grammar> State<'grammar> {
+    pub fn prefix(&self) -> &'grammar [Symbol] {
+        // Each state fn takes as argument the longest prefix of any
+        // item. Note that all items must have compatible prefixes.
+        let (_, prefix) =
+            self.items.vec
+                      .iter()
+                      .map(|item| &item.production.symbols[..item.index])
+                      .map(|symbols| (symbols.len(), symbols))
+                      .max() // grr, max_by is unstable :(
+                      .unwrap();
+
+        debug_assert!(
+            self.items.vec
+                      .iter()
+                      .all(|item| prefix.ends_with(&item.production.symbols[..item.index])));
+
+        prefix
+    }
+}
+
+impl<'grammar> Action<'grammar> {
+    pub fn shift(&self) -> Option<StateIndex> {
+        match *self {
+            Action::Shift(next_index) => Some(next_index),
+            _ => None
+        }
+    }
+    pub fn reduce(&self) -> Option<&'grammar Production> {
+        match *self {
+            Action::Reduce(production) => Some(production),
+            _ => None,
+        }
+    }
+}
