+++ lalrpop/src/lr1/error/mod.rs
@@ -5,10 +5,10 @@ use grammar::repr::*;
 use session::Session;
 use std::io::{self, Result, Write};
 use util::{Map, map};
-use lr1::backtrace::{Example, ExampleStyles, ExampleSymbol, Tracer};
+use lr1::backtrace::Tracer;
+use lr1::core::*;
+use lr1::example::{Example, ExampleStyles, ExampleSymbol};
 use lr1::lookahead::{Lookahead, LookaheadSet};
-use super::{Action, Conflict, LR0Item, Item,
-            State, StateIndex, TableConstructionError};
 
 #[cfg(test)] mod test;
