+++ lalrpop/src/lr1/example/ascii_canvas/row.rs
