+++ lalrpop/src/lr1/interpret.rs
@@ -1,6 +1,6 @@
 //! LR(1) interpeter. Just builds up parse trees. Intended for testing.
 
-use lr1::{Action, State, StateIndex};
+use lr1::core::*;
 use lr1::lookahead::Lookahead;
 use generate::ParseTree;
 use grammar::repr::*;
